modify make hash
