+++ jest-testing-basics/profile.js
@@ -1,4 +1,69 @@
-export default function buildMakeProfile({
+let country = { name: "Uganda", dial_code: "+256", code: "UG" };
+
+let countryCode = country.code;
+
+function makeCountry() {
+  return {
+    getCountryByCountryCode: (code) =>
+      code === countryCode
+        ? country
+        : "The country code provided does not exist.",
+    isValidCountry: (code) => !!(code === countryCode),
+  };
+}
+
+function isPhoneNumberValid(phone) {
+  let parsed = parseInt(phone);
+  if (isNaN(parsed)) {
+    return 0;
+  }
+  return !!parsed;
+}
+
+const ID_TYPES = ["NATIONAL_ID", "DRIVING_PERMIT", "REFUGEE_ID"];
+
+function validateIdentification({ idType, idNumber }) {
+  return !!(
+    typeof idNumber === "string" &&
+    idNumber.length &&
+    ID_TYPES.includes(idType)
+  );
+}
+
+const MALE = "MALE";
+const FEMALE = "FEMALE";
+
+function validateGender(sex) {
+  if (typeof sex !== "string") {
+    return false;
+  }
+  return [MALE, FEMALE].includes(sex);
+}
+
+const MEMBER = "MEMBER";
+const SUPER_ADMIN = "SUPER_ADMIN";
+const ORGANISATION_ADMIN = "ORGANISATION_ADMIN";
+
+function validateUserType(userType) {
+  if (typeof userType !== "string") {
+    return false;
+  }
+  return [MEMBER, SUPER_ADMIN, ORGANISATION_ADMIN].includes(userType);
+}
+
+const create4DigitPIN = () => Math.floor(1000 + Math.random() * 9000);
+
+const makeHash = function (password) {
+  password = "iehncosoan390afoiohajjjajwko2";
+  hashedPassword = password;
+  return hashedPassword;
+};
+
+function makeHashedPassword(passwordToHash) {
+  return makeHash(passwordToHash);
+}
+
+function buildMakeProfile({
   makeCountry,
   isPhoneNumberValid,
   validateIdentification,
@@ -94,3 +159,15 @@ export default function buildMakeProfile({
     });
   };
 }
+
+const makeProfile = buildMakeProfile({
+  makeCountry,
+  isPhoneNumberValid,
+  validateIdentification,
+  validateGender,
+  validateUserType,
+  createRandomPIN: create4DigitPIN,
+  makeHashedPassword,
+});
+
+export default makeProfile;
